Enhance utils (#3)
* FIX: coaches were removed before reconciling with membership list and were consequently being readded during reconciliation

* ENH: new hrefs on site, increase wait time

* ENH utils. REM comments

+++ utilities.py
@@ -13,7 +13,7 @@ from conf_vars import CYCLE, DOWNLOADS_DIR, RESULTS_DIR
 # TODO probably want to change download dir
 
 CHROME_DOWNLOADS = r'C:\Users\cmill\Downloads'
-filenames = glob.glob(rf'{CHROME_DOWNLOADS}\PerformanceResultsWeightlifting*.xlsx')
+lifts = glob.glob(rf'{CHROME_DOWNLOADS}\PerformanceResultsWeightlifting*.xlsx')
 metcons = glob.glob(rf'{CHROME_DOWNLOADS}\PerformanceResultsMetcon*.xlsx')
 
 def setup_dirs():
@@ -39,9 +39,12 @@ def file_rename(filenames, cycle):
         comp = [col.column for col in header if col.value == 'Component'][0]
         liftname = sheet1[comp+'2'].value
         wb.close()
-        dst = f'{DOWNLOADS_DIR}\\{cycle}_{clean_name(liftname)}.xlsx'
-        os.rename(f, dst)
-        print(f'Moved {f} to {dst}')
+        if liftname:
+            dst = f'{DOWNLOADS_DIR}\\{cycle}_{clean_name(liftname)}.xlsx'
+            os.rename(f, dst)
+            print(f'Moved {f} to {dst}')
+        if not liftname:
+            print(f'No data in this file.')
 
 
 def clean_name(liftname):
@@ -74,8 +77,11 @@ def helper_file_rename(cycle):
     for helper in helpers:
         src = rf'{CHROME_DOWNLOADS}\{helper}'
         dst = f'{DOWNLOADS_DIR}\\{cycle}_{helper}'
-        os.rename(src, dst)
-        print(f'Moved {helper} to {DOWNLOADS_DIR}')
+        if os.path.isfile(src):
+            os.rename(src, dst)
+            print(f'Moved {helper} to {DOWNLOADS_DIR}')
+        else:
+            print(f'No such file {helper} at {src}')
 
 
 def main():
@@ -83,7 +89,7 @@ def main():
     """
     setup_dirs()
     helper_file_rename(CYCLE)
-    file_rename(filenames, CYCLE)
+    file_rename(lifts, CYCLE)
     file_rename(metcons, CYCLE)
